requirements.txt
discord.py
requests
pandas
ta